fix: モーダルを1つのコンポーネントで使用できるように修正
