+++ pkg/web/src/pages/browser/[...pathes]/components/ProjectBar.tsx
@@ -2,12 +2,11 @@ import type { BrowserProject, ProjectId } from '@violet/api/types'
 import { alphaLevel, colors, fontSizes } from '@violet/web/src//utils/constants'
 import { pagesPath } from '@violet/web/src/utils/$path'
 import Link from 'next/link'
-import type { ChangeEvent, FormEvent } from 'react'
-import { useCallback, useEffect, useRef, useState } from 'react'
+import { useState } from 'react'
 import styled from 'styled-components'
-import { AddProject } from './AddProject'
+import { InputFromProject } from './InputFromProject'
 import { Modal } from './Modal'
-import { PortalModal } from './PortalModal'
+import { ProjectAddIcon } from './ProjectAddIcon'
 
 const Container = styled.div`
   display: flex;
@@ -49,12 +48,6 @@ const Icon = styled.div`
   border-radius: 6px;
 `
 
-const InputFormProject = styled.div`
-  position: absolute;
-  top: 80px;
-  left: 35%;
-`
-
 const Message = styled.div`
   position: absolute;
   top: 50%;
@@ -65,24 +58,10 @@ const Message = styled.div`
 
 export const ProjectBar = (props: { projects: BrowserProject[]; projectId: ProjectId }) => {
   const [isClickAddProject, setIsClickAddProject] = useState(false)
-  const [label, setLabel] = useState('')
   const [isFocusing, setIsFocusing] = useState(false)
-  const inputLabel = useCallback((e: ChangeEvent<HTMLInputElement>) => setLabel(e.target.value), [])
-  const inputElement = useRef<HTMLInputElement>(null)
-  useEffect(() => {
-    inputElement.current?.focus()
-  }, [inputElement.current, isClickAddProject])
   const addNewProject = () => {
-    if (inputElement && inputElement.current) {
-      inputElement.current.focus()
-    }
     setIsFocusing(false)
     setIsClickAddProject(true)
-  }
-  const sendProjectName = (e: FormEvent) => {
-    e.preventDefault()
-    setIsFocusing(!label)
-    setLabel('')
   }
 
   const closeModal = () => {
@@ -102,23 +81,12 @@ export const ProjectBar = (props: { projects: BrowserProject[]; projectId: Proje
           </IconWrapper>
         </Link>
       ))}
-      <AddProject addProject={addNewProject} />
+      <ProjectAddIcon addProject={addNewProject} />
       {isClickAddProject && !isFocusing && (
-        <PortalModal>
-          <Modal closeModal={closeModal}>
-            <Message>{'Please enter a project name'}</Message>
-            <InputFormProject>
-              <form onSubmit={sendProjectName}>
-                <input
-                  ref={inputElement}
-                  type="text"
-                  onBlur={sendProjectName}
-                  onChange={inputLabel}
-                />
-              </form>
-            </InputFormProject>
-          </Modal>
-        </PortalModal>
+        <Modal closeModal={closeModal}>
+          <Message>{'Please enter a project name'}</Message>
+          <InputFromProject closeModal={closeModal} />
+        </Modal>
       )}
     </Container>
   )
